@@ -7,48 +7,56 @@ by sending a chat completion request and receiving a response without
 using any mock objects.
 """
 
-import subprocess
-import time
-import requests
-import json
-import sys
-import os
 import concurrent.futures
-from src.proxy.config import Config
+import json
+import os
+import subprocess
+import sys
+import time
+from typing import Optional
+
+import ollama
+import requests
+
+from src.frameworks_drivers.config import Config
+from tests.backend_checker_factory import BackendCheckerFactory
+
+
 class E2ESimulation:
     """Handles end-to-end simulation of the Llama Smart Proxy."""
+
     def __init__(self):
 
+        # Load original configuration data
+        with open("config.json") as f:
+            self.original_config_data = json.load(f)
 
         # Load configuration
-        self.config = Config.load('config.json')
+        self.config = Config.load("config.json")
         if self.config.simulation is None:
             raise ValueError("Simulation configuration is required for this script")
         self.sim = self.config.simulation
         if self.sim.enable_remote:
             host = self.sim.host
-            if host.startswith('http://'):
+            if host.startswith("http://"):
                 host = host[7:]
-            elif host.startswith('https://'):
+            elif host.startswith("https://"):
                 host = host[8:]
             self.server_url = f"http://{host}:{self.sim.port}"
         else:
             self.server_url = self.sim.server_url or f"http://localhost:{self.config.server.port}"
 
+        # Create backend checker
+        self.checker_factory = BackendCheckerFactory(self.config.model_dump())
+        self.checker = self.checker_factory.create_checker()
+
     def get_base_request_data(self):
         """Get the base request data for chat completion."""
-        return {
-            "model": self.sim.model,
-            "messages": [msg.dict() for msg in self.sim.messages]
-        }
+        return {"model": self.sim.model, "messages": [msg.model_dump() for msg in self.sim.messages]}
 
     def send_request(self, endpoint, data):
         """Send a POST request to the specified endpoint with the given data."""
-        return requests.post(
-            self.server_url + endpoint,
-            json=data,
-            timeout=self.sim.request_timeout
-        )
+        return requests.post(self.server_url + endpoint, json=data, timeout=self.sim.request_timeout)
 
     def wait_for_server(self, url: str, timeout: int = 30):
         """Wait for the server to be ready by checking the health endpoint."""
@@ -62,11 +70,17 @@ class E2ESimulation:
             time.sleep(1)
         return None
 
-    def run(self):
+
+    def run_single_simulation(self):
         """Main simulation function."""
 
         result = 1
         print("Starting Llama Smart Proxy E2E Simulation...")
+
+        # Check backend availability
+        if not self.checker.check_availability():
+            print(f"{self.config.backend} backend not ready. Skipping simulation.")
+            return result
 
         server_process = None
         if not self.sim.enable_remote:
@@ -77,7 +91,8 @@ class E2ESimulation:
 
         try:
             # Run all checks
-            checks = [self.check_completion, self.check_parallel_requests, self.check_forwarding_endpoints]
+            # , self.check_completion, self.check_parallel_requests
+            checks = [self.check_forwarding_endpoints, self.check_parallel_requests]
             for check in checks:
                 if not check():
                     return result
@@ -96,14 +111,14 @@ class E2ESimulation:
 
             # Server output is printed directly
 
-        print("E2E simulation completed successfully!")
+        print("Simulation completed!")
         return result
 
     def check_completion(self):
         result = False
         request_data = self.get_base_request_data()
 
-        # Send the request
+        # Send the request via HTTP
         response = self.send_request(self.sim.chat_endpoint, request_data)
 
         if response.status_code == 200:
@@ -123,21 +138,19 @@ class E2ESimulation:
         print("Testing parallel requests...")
 
         request_data = self.get_base_request_data()
-        request_data["max_tokens"] = 512
+        request_data["max_tokens"] = 2048
 
         def send_request():
             try:
                 response = self.send_request(self.sim.chat_endpoint, request_data)
                 if response.status_code == 200:
                     result = response.json()
-                    print(f'Response: {result}')
+                    print(f"Response: {result}")
                     # Check if response has choices with content
-                    if 'choices' in result and len(result['choices']) > 0 and 'message' in result['choices'][0]:
+                    if "choices" in result and len(result["choices"]) > 0 and "message" in result["choices"][0]:
                         return True
-                    else:
-                        return False
-                else:
                     return False
+                return False
             except requests.RequestException:
                 return False
 
@@ -149,29 +162,24 @@ class E2ESimulation:
         if success_count == 4:
             print("All parallel requests successful!")
             return True
-        else:
-            print(f"Parallel requests failed: {success_count}/4 successful")
-            return False
+        print(f"Parallel requests failed: {success_count}/4 successful")
+        return False
 
     def check_forwarding_endpoints(self):
         """Test forwarding of endpoints that are not processed by the proxy."""
         print("Testing forwarding endpoints...")
         print("Note: Forwarding requires a running backend server")
 
-        forwarding_endpoints = [
-            "/tokenize",
-            "/detokenize",
-            # "/embedding",
-            # "/props",
-            # "/reranking"
-        ]
+        forwarding_endpoints = self.checker.get_forwarding_endpoints()
 
         backend_unavailable_count = 0
         success_count = 0
-        for endpoint in forwarding_endpoints:
+        for endpoint_info in forwarding_endpoints:
+            endpoint = endpoint_info["endpoint"]
+            data_key = endpoint_info["data_key"]
             try:
-                response = self.send_request(endpoint, {"model": self.sim.model})
-                # print(f"Response: {response.json()}")
+                data = {data_key: self.sim.model}
+                response = self.send_request(endpoint, data)
                 if response.status_code == 200:
                     print(f"✓ {endpoint} forwarded successfully")
                     success_count += 1
@@ -186,15 +194,16 @@ class E2ESimulation:
         if success_count == len(forwarding_endpoints):
             print("All forwarding endpoints tested successfully")
             return True
-        elif backend_unavailable_count == len(forwarding_endpoints):
+        if backend_unavailable_count == len(forwarding_endpoints):
             print("All forwarding endpoints report backend unavailable (expected when no backend server is running)")
             return True  # Don't fail the simulation if backend is not running
-        elif success_count + backend_unavailable_count == len(forwarding_endpoints):
-            print(f"Forwarding test completed: {success_count} successful, {backend_unavailable_count} backend unavailable")
+        if success_count + backend_unavailable_count == len(forwarding_endpoints):
+            print(
+                f"Forwarding test completed: {success_count} successful, {backend_unavailable_count} backend unavailable",
+            )
             return True
-        else:
-            print(f"Partial success: {success_count}/{len(forwarding_endpoints)} endpoints forwarded")
-            return True
+        print(f"Partial success: {success_count}/{len(forwarding_endpoints)} endpoints forwarded")
+        return True
 
     def stop_server(self, server_process):
         print("Stopping server...")
@@ -208,24 +217,70 @@ class E2ESimulation:
     def start_server(self):
         try:
             print("Starting proxy server...")
-            server_process = subprocess.Popen(
-                [sys.executable, 'main.py'],
-                cwd=os.getcwd()
-            )
+            server_process = subprocess.Popen([sys.executable, "main.py"], cwd=os.getcwd())
             # Wait for server to be ready
             print("Waiting for server to start...")
-            check_response = self.wait_for_server(self.server_url + self.sim.health_endpoint, timeout=self.sim.wait_timeout)
+            check_response = self.wait_for_server(
+                self.server_url + self.sim.health_endpoint, timeout=self.sim.wait_timeout,
+            )
             if not check_response:
                 print("Server failed to start within timeout")
                 return None
 
-            print(f"Server is ready. Sending chat completion request...")
+            print("Server is ready. Sending chat completion request...")
 
         except Exception as e:
             print(f"Failed to start server: {e}")
             server_process = None
-            
+
         return server_process
+
+    def run(self):
+        # 
+        backends = ["ollama", "llama.cpp"]
+
+        overall_result = 0
+
+        for backend in backends:
+            print(f"\n{'='*50}")
+            print(f"Testing backend: {backend}")
+            print(f"{'='*50}")
+
+            # Modify config
+            config_data = self.original_config_data.copy()
+            config_data["backend"] = backend
+
+            with open("config.json", "w") as f:
+                json.dump(config_data, f, indent=2)
+
+            # Reload config
+            self.config = Config(**config_data)
+
+            # Update checker factory config and recreate checker
+            self.checker_factory.config = self.config.model_dump()
+            self.checker = self.checker_factory.create_checker()
+
+            # Set the simulation model using the backend checker
+            config_data["simulation"]["model"] = self.checker.get_simulation_model()
+
+            with open("config.json", "w") as f:
+                json.dump(config_data, f, indent=2)
+
+            # Reload config again
+            self.config = Config(**config_data)
+            self.sim = self.config.simulation
+
+            # Run simulation for this backend
+            result = self.run_single_simulation()
+            if result != 0:
+                overall_result = result
+
+        # Restore original config
+        with open("config.json", "w") as f:
+            json.dump(self.original_config_data, f, indent=2)
+
+        print("E2E simulation completed for all backends!")
+        return overall_result
 
 
 if __name__ == "__main__":
