@@ -4,6 +4,7 @@ class ModelResolver:
     """
 
     DEFAULT_VARIANT = "Q4_K_M"
+    _cache: dict[str, tuple[str, str]] = {}
 
     @staticmethod
     def resolve(model_identifier: str) -> tuple[str, str]:
@@ -23,26 +24,31 @@ class ModelResolver:
         Raises:
             ValueError: If the model identifier is invalid.
         """
+        if model_identifier in ModelResolver._cache:
+            return ModelResolver._cache[model_identifier]
+
         if not model_identifier or not model_identifier.strip():
             raise ValueError("Model identifier cannot be empty")
 
         model_identifier = model_identifier.strip()
 
-        if ':' in model_identifier:
-            parts = model_identifier.split(':', 1)
+        if ":" in model_identifier:
+            parts = model_identifier.split(":", 1)
             if len(parts) != 2:
                 raise ValueError(f"Invalid model identifier format: {model_identifier}")
             repo_id, variant = parts
-            if not repo_id or not variant or ':' in variant:
+            if not repo_id or not variant or ":" in variant:
                 raise ValueError(f"Invalid model identifier format: {model_identifier}")
         else:
             repo_id = model_identifier
             variant = ModelResolver.DEFAULT_VARIANT
 
         # Create filename pattern for GGUF files
-        if variant.endswith('.gguf'):
+        if variant.endswith(".gguf"):
             filename_pattern = variant
         else:
             filename_pattern = f"*{variant}.gguf"
 
-        return repo_id, filename_pattern
+        result = repo_id, filename_pattern
+        ModelResolver._cache[model_identifier] = result
+        return result
